refactor: 121-move files into slider folder
