+++ src/components/product/slider/slider.css
